requirements.txt
ultralytics
opencv-python
PyYAML
Pillow
tqdm
